+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 7. **Add feature**: Session restore now includes cursor position and text selection state. Your cursor will be exactly where you left it!
 8. **Add feature**: *(Now planned)* Add **project** feature
 9. **Fix issue**: Fix issue that it will make a crash if you make a type=0 plugin and put it to `./resources/plugins/`
+10. **Documentation Edit**: Edit `function.md`, add Q&A chapter.
 ### 0.06.25542
 1. **Add feature**: Add **+** button to create new Tab. (I forgot add it in 0.06.25530)
 2. **Add feature**: It automatically restores last session every start. (It won't be restored your cursor position.)